+++ CHANGELOG.md
@@ -3,18 +3,24 @@
 # 1.2.0 (xxxx-xx-xx)
 
 * Gameplay:
+    * Player can now SPEED SURGE (left trigger / X key) to help avoid enemies.
     * Upgrades:
         * Player now has a choice of 3 between waves (was: 2)
         * NEW: HEAL ME - Restore all health ahead of the next wave.
-        * NEW: SPEED SURGE - Left trigger gives the player a SURGE of SPEED.
+        * NEW: SPEED SURGE - Improves SPEED SURGE recharge time.
     * Chasing enemies now use proper navigation instead of hoping for the best.  Beware! 
+* Misc: 
+    * Graphics: 'DINNER TIME' particle effects added.
+* Platform:
+    * STEAM ACHIEVEMENTS!
+        * Beat 1 wave.
 * Bug fixes:
     * Deadlock could occur on upgrade screen if there weren't enough eligible upgrades. 
     * Player could get stuck hunting for key or exit.  Proper navigation implemented to avoid this.
     * Enemies vanishing too early when wave ending during SHARK ATTACK.
     * Shark spray sound synced properly with gameplay.
     * Game ending when in FISH FRENZY would result in player stuck at a weird angle.
-    * Player and Mini Shark projectiles now despawn on game ending.
+    * Player, Mini Shark and Dinosaur projectiles now despawn on game ending.
 * Misc:
     * [MAYBE] Zoom camera effects for wave start and end.
     * Adjusted player movement at wave start to commence at the entrance door.
@@ -42,7 +48,7 @@
     * Diet: At some point, Necromancers stopped wanting to eat fish.  Fixed.
     * UI: Tightened up mouse showing / hiding as appropriate
 * Platform:
-    * Steam : GodotSteam GDExtension upgraded to 4.6.
+    * Steam: GodotSteam GDExtension upgraded to 4.6.
 * Misc:
     * Haptic feedback added! Vibrates on player hit + shark frenzy.  Turn on in OPTIONS.
         * This will not work on all platforms.